packages/jb-storage/pyproject.toml
[tool.poetry]
name = "jb-storage"
version = "0.1.0"
description = ""
authors = ["OpenNyAI Team <opennyai@googlegroups.com>"]
readme = "README.md"
packages = [{include = "jugalbandi/storage"}]

[tool.poetry.dependencies]
python = ">=3.10, <4.0.0"
gcloud-aio-storage = "^8.2.0"
google-cloud-storage = "^2.9.0"
tenacity = "^8.2.2"
aiohttp = "^3.8.4"
aiofiles = "^23.1.0"
types-aiofiles = "^23.1.0.4"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"
poethepoet = "^0.20.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
faker = "^18.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
